Layer seperation is implemented (Closes #12)
