+++ lib/core/supabase/supabase_repository.dart
@@ -1,4 +1,5 @@
-import 'package:eunoia_chat_application/features/chat/data/models/conversation_model.dart';
+import 'package:eunoia_chat_application/core/shared_preferences/shared_preferences_user_manager.dart';
+import 'package:eunoia_chat_application/features/conversation/data/models/conversation_model.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:realtime_client/realtime_client.dart';
 
@@ -10,11 +11,17 @@ abstract class SupabaseRepository {
           callBackFunc}) async {
     final channel = socket.channel('realtime:public:chat');
     ConversationModel? conversationModel;
+    print(((await SharedPreferencesUserManager.getUser())?.user.id));
     channel
         .onPostgresChanges(
           event: PostgresChangeEvent.all,
           table: 'conversation',
           schema: 'public',
+          filter: PostgresChangeFilter(
+            type: PostgresChangeFilterType.eq,
+            column: 'creator_id',
+            value: ((await SharedPreferencesUserManager.getUser())?.user.id),
+          ),
           callback: (payload) {
             conversationModel = ConversationModel.fromJson(payload.newRecord);
             if (conversationModel != null) {
