+++ lib/features/message/data/datasources/message_remote_data_source.dart
@@ -1,0 +1,1 @@
+class MessageRemoteDataSource {}
