+++ lib/features/message/data/datasources/message_remote_data_source_impl.dart
@@ -1,0 +1,3 @@
+import 'package:eunoia_chat_application/features/message/data/datasources/message_remote_data_source.dart';
+
+class MessageRemoteDataSourceImpl implements MessageRemoteDataSource {}
