+++ lib/features/message/data/repositories/message_repository_impl.dart
@@ -1,0 +1,3 @@
+import 'package:eunoia_chat_application/features/message/domain/repositories/message_repository.dart';
+
+class MessageRepositoryImpl implements MessageRepository {}
