+++ lib/features/message/domain/entities/message.dart
