+++ lib/features/message/domain/entities/participant_type.dart
