+++ lib/features/message/domain/repositories/message_repository.dart
@@ -1,0 +1,1 @@
+abstract class MessageRepository {}
