+++ lib/features/message/presentation/cubit/message_cubit.dart
@@ -1,0 +1,8 @@
+import 'package:equatable/equatable.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+
+part 'message_state.dart';
+
+class MessageCubit extends Cubit<MessageState> {
+  MessageCubit() : super(MessageInitial());
+}
