+++ lib/features/message/presentation/cubit/message_state.dart
@@ -1,0 +1,10 @@
+part of 'message_cubit.dart';
+
+abstract class MessageState extends Equatable {
+  const MessageState();
+
+  @override
+  List<Object> get props => [];
+}
+
+class MessageInitial extends MessageState {}
