+++ lib/injection.dart
@@ -1,14 +1,19 @@
 import 'package:dio/dio.dart';
 import 'package:eunoia_chat_application/core/dio/interceptor.dart';
 import 'package:eunoia_chat_application/features/authentication/presentation/cubit/authentication_cubit.dart';
-import 'package:eunoia_chat_application/features/chat/data/datasources/chat_remote_data_source.dart';
-import 'package:eunoia_chat_application/features/chat/data/datasources/chat_remote_data_source_impl.dart';
-import 'package:eunoia_chat_application/features/chat/data/repositories/chat_repository_impl.dart';
-import 'package:eunoia_chat_application/features/chat/domain/repositories/chat_repository.dart';
-import 'package:eunoia_chat_application/features/chat/domain/usecases/get_conversations_usecase.dart';
-import 'package:eunoia_chat_application/features/chat/domain/usecases/listen_conversations_usecase.dart';
-import 'package:eunoia_chat_application/features/chat/presentation/cubit/chat_cubit.dart';
+import 'package:eunoia_chat_application/features/conversation/data/datasources/conversation_remote_data_source.dart';
+import 'package:eunoia_chat_application/features/conversation/data/datasources/conversation_remote_data_source_impl.dart';
+import 'package:eunoia_chat_application/features/conversation/data/repositories/conversation_repository_impl.dart';
+import 'package:eunoia_chat_application/features/conversation/domain/repositories/conversation_repository.dart';
+import 'package:eunoia_chat_application/features/conversation/domain/usecases/get_conversations_usecase.dart';
+import 'package:eunoia_chat_application/features/conversation/domain/usecases/listen_conversations_usecase.dart';
+import 'package:eunoia_chat_application/features/conversation/presentation/cubit/conversation_cubit.dart';
 import 'package:eunoia_chat_application/features/main/presentation/cubit/main_cubit.dart';
+import 'package:eunoia_chat_application/features/message/data/datasources/message_remote_data_source.dart';
+import 'package:eunoia_chat_application/features/message/data/datasources/message_remote_data_source_impl.dart';
+import 'package:eunoia_chat_application/features/message/data/repositories/message_repository_impl.dart';
+import 'package:eunoia_chat_application/features/message/domain/repositories/message_repository.dart';
+import 'package:eunoia_chat_application/features/message/presentation/cubit/message_cubit.dart';
 import 'package:eunoia_chat_application/features/user/data/datasources/user_remote_data_source.dart';
 import 'package:eunoia_chat_application/features/user/data/datasources/user_remote_data_source_impl.dart';
 import 'package:eunoia_chat_application/features/user/data/repositories/user_repository_impl.dart';
@@ -34,8 +39,13 @@ initCubits() {
   getIt.registerFactory(
       () => UserCubit(userLoginUsecase: getIt(), userRegisterUsecase: getIt()));
   getIt.registerFactory(() => MainCubit());
-  getIt.registerFactory(() =>
-      ChatCubit(getConversationsUsecase: getIt(), listenConversationsUsecase: getIt()));
+  getIt.registerFactory(() => MessageCubit());
+
+  getIt.registerFactory(() => ConversationCubit(
+        getConversationsUsecase: getIt(),
+        listenConversationsUsecase: getIt(),
+      ));
+
   getIt.registerFactory(() => AuthenticationCubit());
 }
 
@@ -43,8 +53,12 @@ initDataSources() {
   getIt.registerLazySingleton<UserRemoteDataSource>(
     () => UserRemoteDataSourceImpl(),
   );
-  getIt.registerLazySingleton<ChatRemoteDataSource>(
-    () => ChatRemoteDataSourceImpl(),
+
+  getIt.registerLazySingleton<ConversationRemoteDataSource>(
+    () => ConversationRemoteDataSourceImpl(),
+  );
+  getIt.registerLazySingleton<MessageRemoteDataSource>(
+    () => MessageRemoteDataSourceImpl(),
   );
 }
 
@@ -59,8 +73,12 @@ initRepositories() {
   getIt.registerLazySingleton<UserRepository>(
     () => UserRepositoryImpl(userRemoteDataSource: getIt()),
   );
-  getIt.registerLazySingleton<ChatRepository>(
-    () => ChatRepositoryImpl(remoteDataSource: getIt()),
+
+  getIt.registerLazySingleton<ConversationRepository>(
+    () => ConversationRepositoryImpl(remoteDataSource: getIt()),
+  );
+  getIt.registerLazySingleton<MessageRepository>(
+    () => MessageRepositoryImpl(),
   );
 }
